Cdscrs 33029 config be (#206)
* initial

* still not working

* meglio

* it works

* cleanup
