+++ pkg/plugin/query.go
@@ -25,7 +25,13 @@ func (ocidx *OCIDatasource) query(ctx context.Context, pCtx backend.PluginContex
 
 	// checking if the query has valid tenancy detail
 	if qm.TenancyOCID == "" {
-		//backend.Logger.Warn("plugin.query", "query", "tenancy ocid is mandatory but it is not present in query")
+		backend.Logger.Warn("plugin.query", "query", "tenancy ocid is mandatory but it is not present in query")
+		return response
+	}
+
+	// checking if the query has valid Interval detail
+	if qm.Interval == "" {
+		backend.Logger.Warn("plugin.query", "query", "Interval is mandatory but it is not present in query")
 		return response
 	}
