+++ src/datasource.ts
@@ -43,14 +43,9 @@ export class OCIDataSource extends DataSourceWithBackend<OCIQuery, OCIDataSource
   }
 
 
-  async getTenancyMode(): Promise<string[]> {
-    return this.getResource(OCIResourceCall.Tenancies).then((response) => {
-      return new ResponseParser().parseTenancyMode(response);
-    });
-  }
-
   async getTenancies(): Promise<OCIResourceItem[]> {
     return this.getResource(OCIResourceCall.Tenancies).then((response) => {
+      console.log ("Ritorno di Tenanci");
       return new ResponseParser().parseTenancies(response);
     });
   }
